# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FireISP 5.0 API",
    "version": "5.0.0",
    "description": "Open source ISP management — customers, plans, billing, network monitoring, and Mexican fiscal compliance (CFDI 4.0).",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "/api",
      "description": "FireISP API"
    }
  ],
...
  "paths": {
...
    "/export/invoices": {
      "get": {
        "tags": [
          "Export"
        ],
        "summary": "Export invoices as CSV",
        "operationId": "exportInvoices",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "File download",
            "content": {
              "text/csv": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/export/clients": {
      "get": {
        "tags": [
          "Export"
        ],
        "summary": "Export clients as CSV",
        "operationId": "exportClients",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "File download",
            "content": {
              "text/csv": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/export/contracts": {
      "get": {
        "tags": [
          "Export"
        ],
        "summary": "Export contracts as CSV",
        "operationId": "exportContracts",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "File download",
            "content": {
              "text/csv": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/export/payments": {
      "get": {
        "tags": [
          "Export"
        ],
        "summary": "Export payments as CSV",
        "operationId": "exportPayments",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "File download",
            "content": {
              "text/csv": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/import/clients": {
      "post": {
        "tags": [
          "Import"
        ],
        "summary": "Bulk import clients from CSV",
        "operationId": "importClients",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "description": "csv",
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "ImportResult",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/import/devices": {
      "post": {
        "tags": [
          "Import"
        ],
        "summary": "Bulk import devices from CSV",
        "operationId": "importDevices",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "description": "csv",
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "ImportResult",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/import/contracts": {
      "post": {
        "tags": [
          "Import"
        ],
        "summary": "Bulk import contracts from CSV",
        "operationId": "importContracts",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "description": "csv",
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "ImportResult",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  }
}